+++ lib/puppet/provider/cffirehol_config/firehol.rb
@@ -21,8 +21,6 @@ Puppet::Type.type(:cffirehol_config).provide(
             :name => 'firehol',
             :ensure => fhmeta['generator_version'].empty? ? :absent : :exists,
             :custom_headers => fhmeta['custom_headers'],
-            :ip_whitelist => fhmeta['ip_whitelist'],
-            :ip_blacklist => fhmeta['ip_blacklist'],
             :synproxy_public => fhmeta['synproxy_public'],
             :persistent_dhcp => fhmeta['persistent_dhcp'],
             :enable => fhmeta['enable'],
@@ -39,15 +37,11 @@ Puppet::Type.type(:cffirehol_config).provide(
         case ensure_val 
         when :absent
             write_config('custom_headers', [])
-            write_config('ip_whitelist', [])
-            write_config('ip_blacklist', [])
             write_config('synproxy_public', false)
             write_config('persistent_dhcp', false)
             write_config('enable', false)
         when :present, :exists
             write_config('custom_headers', (@resource[:custom_headers] or []))
-            write_config('ip_whitelist', (@resource[:ip_whitelist] or []))
-            write_config('ip_blacklist', (@resource[:ip_blacklist] or []))
             write_config('synproxy_public', (@resource[:synproxy_public] or false))
             write_config('persistent_dhcp', (@resource[:persistent_dhcp] or false))
             write_config('enable', (@resource[:enable] or false))